Moved both dice.js and chat.js to helper folder
