+++ module/SBB.mjs
@@ -5,7 +5,7 @@ import { SBBItemSheet } from "./sheets/item-sheet.mjs";
 // Import helper/utility classes and constants.
 import { preloadHandlebarsTemplates } from "./helpers/templates.mjs";
 import { SBB } from "./helpers/config.mjs";
-import * as Chat from "./chat/chat.mjs";
+import * as Chat from "./helpers/chat.mjs";
 
 Hooks.once("init", function (){
     console.log("SBB | Loading Stars but Butter");
